+++ usvm-core/src/main/kotlin/org/usvm/collection/array/UArrayModelRegion.kt
@@ -1,13 +1,8 @@
 package org.usvm.collection.array
 
 import io.ksmt.solver.KModel
-import org.usvm.UBoolExpr
-import org.usvm.UConcreteHeapAddress
 import org.usvm.UExpr
-import org.usvm.UHeapRef
-import org.usvm.USizeExpr
 import org.usvm.USort
-import org.usvm.memory.UMemoryRegion
 import org.usvm.memory.UReadOnlyMemoryRegion
 import org.usvm.model.AddressesMapping
 import org.usvm.model.modelEnsureConcreteInputRef
@@ -16,8 +11,7 @@ import org.usvm.solver.UCollectionDecoder
 
 abstract class UArrayModelRegion<ArrayType, Sort : USort>(
     private val regionId: UArrayRegionId<ArrayType, Sort>,
-) : UArrayRegion<ArrayType, Sort> {
-
+) : UReadOnlyMemoryRegion<UArrayIndexLValue<ArrayType, Sort>, Sort> {
     val defaultValue by lazy { regionId.sort.sampleUValue() }
 
     abstract val inputArray: UReadOnlyMemoryRegion<USymbolicArrayIndex, Sort>?
@@ -25,37 +19,6 @@ abstract class UArrayModelRegion<ArrayType, Sort : USort>(
     override fun read(key: UArrayIndexLValue<ArrayType, Sort>): UExpr<Sort> {
         val ref = modelEnsureConcreteInputRef(key.ref) ?: return defaultValue
         return inputArray?.read(ref to key.index) ?: defaultValue
-    }
-
-    override fun write(
-        key: UArrayIndexLValue<ArrayType, Sort>,
-        value: UExpr<Sort>,
-        guard: UBoolExpr
-    ): UMemoryRegion<UArrayIndexLValue<ArrayType, Sort>, Sort> {
-        error("Illegal operation for a model")
-    }
-
-    override fun memcpy(
-        srcRef: UHeapRef,
-        dstRef: UHeapRef,
-        type: ArrayType,
-        elementSort: Sort,
-        fromSrcIdx: USizeExpr,
-        fromDstIdx: USizeExpr,
-        toDstIdx: USizeExpr,
-        guard: UBoolExpr
-    ): UArrayRegion<ArrayType, Sort> {
-        error("Illegal operation for a model")
-    }
-
-    override fun initializeAllocatedArray(
-        address: UConcreteHeapAddress,
-        arrayType: ArrayType,
-        sort: Sort,
-        content: Map<USizeExpr, UExpr<Sort>>,
-        guard: UBoolExpr
-    ): UArrayRegion<ArrayType, Sort> {
-        error("Illegal operation for a model")
     }
 }
