+++ usvm-core/src/main/kotlin/org/usvm/collection/array/length/UArrayLengthModelRegion.kt
@@ -1,11 +1,9 @@
 package org.usvm.collection.array.length
 
 import io.ksmt.solver.KModel
-import org.usvm.UBoolExpr
 import org.usvm.UExpr
 import org.usvm.UHeapRef
 import org.usvm.USizeSort
-import org.usvm.memory.UMemoryRegion
 import org.usvm.memory.UReadOnlyMemoryRegion
 import org.usvm.model.AddressesMapping
 import org.usvm.model.modelEnsureConcreteInputRef
@@ -14,7 +12,7 @@ import org.usvm.solver.UCollectionDecoder
 
 abstract class UArrayLengthModelRegion<ArrayType>(
     private val regionId: UArrayLengthsRegionId<ArrayType>,
-) : UArrayLengthsRegion<ArrayType> {
+) : UReadOnlyMemoryRegion<UArrayLengthLValue<ArrayType>, USizeSort> {
     val defaultValue by lazy { regionId.sort.sampleUValue() }
 
     abstract val inputArrayLength: UReadOnlyMemoryRegion<UHeapRef, USizeSort>?
@@ -22,14 +20,6 @@ abstract class UArrayLengthModelRegion<ArrayType>(
     override fun read(key: UArrayLengthLValue<ArrayType>): UExpr<USizeSort> {
         val ref = modelEnsureConcreteInputRef(key.ref) ?: return defaultValue
         return inputArrayLength?.read(ref) ?: defaultValue
-    }
-
-    override fun write(
-        key: UArrayLengthLValue<ArrayType>,
-        value: UExpr<USizeSort>,
-        guard: UBoolExpr
-    ): UMemoryRegion<UArrayLengthLValue<ArrayType>, USizeSort> {
-        error("Illegal operation for a model")
     }
 }
