+++ usvm-core/src/main/kotlin/org/usvm/collection/map/length/UMapLengthModelRegion.kt
@@ -1,11 +1,9 @@
 package org.usvm.collection.map.length
 
 import io.ksmt.solver.KModel
-import org.usvm.UBoolExpr
 import org.usvm.UExpr
 import org.usvm.UHeapRef
 import org.usvm.USizeSort
-import org.usvm.memory.UMemoryRegion
 import org.usvm.memory.UReadOnlyMemoryRegion
 import org.usvm.model.AddressesMapping
 import org.usvm.model.modelEnsureConcreteInputRef
@@ -14,7 +12,7 @@ import org.usvm.solver.UCollectionDecoder
 
 abstract class UMapLengthModelRegion<MapType>(
     private val regionId: UMapLengthRegionId<MapType>,
-) : UMapLengthRegion<MapType> {
+) : UReadOnlyMemoryRegion<UMapLengthLValue<MapType>, USizeSort> {
     val defaultValue by lazy { regionId.sort.sampleUValue() }
 
     abstract val inputMapLength: UReadOnlyMemoryRegion<UHeapRef, USizeSort>?
@@ -22,14 +20,6 @@ abstract class UMapLengthModelRegion<MapType>(
     override fun read(key: UMapLengthLValue<MapType>): UExpr<USizeSort> {
         val ref = modelEnsureConcreteInputRef(key.ref) ?: return defaultValue
         return inputMapLength?.read(ref) ?: defaultValue
-    }
-
-    override fun write(
-        key: UMapLengthLValue<MapType>,
-        value: UExpr<USizeSort>,
-        guard: UBoolExpr
-    ): UMemoryRegion<UMapLengthLValue<MapType>, USizeSort> {
-        error("Illegal operation for a model")
     }
 }
