+++ usvm-core/src/main/kotlin/org/usvm/collection/map/primitive/UMapModelRegion.kt
@@ -1,11 +1,9 @@
 package org.usvm.collection.map.primitive
 
 import io.ksmt.solver.KModel
-import org.usvm.UBoolExpr
 import org.usvm.UExpr
 import org.usvm.USort
 import org.usvm.collection.map.USymbolicMapKey
-import org.usvm.memory.UMemoryRegion
 import org.usvm.memory.UReadOnlyMemoryRegion
 import org.usvm.model.AddressesMapping
 import org.usvm.model.modelEnsureConcreteInputRef
@@ -15,7 +13,7 @@ import org.usvm.util.Region
 
 abstract class UMapModelRegion<MapType, KeySort : USort, ValueSort : USort, Reg : Region<Reg>>(
     private val regionId: UMapRegionId<MapType, KeySort, ValueSort, Reg>
-) : UMapRegion<MapType, KeySort, ValueSort, Reg> {
+) : UReadOnlyMemoryRegion<UMapEntryLValue<MapType, KeySort, ValueSort, Reg>, ValueSort> {
     val defaultValue by lazy { regionId.sort.sampleUValue() }
 
     abstract val inputMap: UReadOnlyMemoryRegion<USymbolicMapKey<KeySort>, ValueSort>?
@@ -23,14 +21,6 @@ abstract class UMapModelRegion<MapType, KeySort : USort, ValueSort : USort, Reg 
     override fun read(key: UMapEntryLValue<MapType, KeySort, ValueSort, Reg>): UExpr<ValueSort> {
         val mapRef = modelEnsureConcreteInputRef(key.mapRef) ?: return defaultValue
         return inputMap?.read(mapRef to key.mapKey) ?: defaultValue
-    }
-
-    override fun write(
-        key: UMapEntryLValue<MapType, KeySort, ValueSort, Reg>,
-        value: UExpr<ValueSort>,
-        guard: UBoolExpr
-    ): UMemoryRegion<UMapEntryLValue<MapType, KeySort, ValueSort, Reg>, ValueSort> {
-        error("Illegal operation for a model")
     }
 }
