+++ usvm-core/src/test/kotlin/org/usvm/api/collections/ObjectMapTest.kt
@@ -4,13 +4,13 @@ import io.ksmt.solver.KSolver
 import org.junit.jupiter.api.Disabled
 import org.usvm.UHeapRef
 import org.usvm.USizeExpr
-import org.usvm.api.ObjectMapCollectionApi.mkSymbolicObjectMap
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapContains
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapGet
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapMergeInto
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapPut
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapRemove
-import org.usvm.api.ObjectMapCollectionApi.symbolicObjectMapSize
+import org.usvm.api.collection.ObjectMapCollectionApi.mkSymbolicObjectMap
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapContains
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapGet
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapMergeInto
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapPut
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapRemove
+import org.usvm.api.collection.ObjectMapCollectionApi.symbolicObjectMapSize
 import org.usvm.model.UModelBase
 import org.usvm.solver.USatResult
 import org.usvm.types.single.SingleTypeSystem
